requirements.txt
bumpversion
flake8>=2.2.3
pytest-cov
pytest>=3.9
sh>=1.09
tox
tox-interpreters
wheel
twine
python-dotenv
pyAesCrypt
